change quantity

+++ model/ShoppingCartModel.php
@@ -30,14 +30,14 @@ class ShoppingCartModel {
         $sum = 0;
         for($i = 0; $i < count($listShoppingCarts); $i++) {
           $index = $listShoppingCarts[$i]["phonecaseid"];
-          $consulta3 = $this->connection->prepare("SELECT name,price,value,description, image FROM phonecase where id = $index");
+          $consulta3 = $this->connection->prepare("SELECT name,price,value,description, image,id FROM phonecase where id = $index");
           $consulta3->execute();
           $resultados3 = $consulta3 ->fetchAll();
           $cost =  $resultados3[0]["price"] * $listShoppingCarts[$i]["amount"];
           $indexResutl3 = count($resultados3[0]);
           $resultados3[0]["cost"] = $cost;
           $resultados3[0]["cartId"] = $listShoppingCarts[$i]["id"];
-          $resultados3[0]["quantity"] = $listShoppingCarts[$i][1];
+          $resultados3[0]["quantity"] = $listShoppingCarts[$i]["amount"];
           array_push($listProducts,$resultados3[0]);
           $sum = $sum + $cost;
       }
@@ -52,17 +52,49 @@ class ShoppingCartModel {
      
     }
     public function deleteCart($id){
-      $sql = "DELETE FROM shoppingcart WHERE id=$id";
-      $consulta = $this->connection->prepare($sql);
-      $consulta->execute();
-      if($consulta == true){
-          echo '<script>alert("Delete thành công")</script>';
-          require_once __DIR__ . "/../view/shoppingCart.php"; 
-      }
-      else{
+      try{
+        $quantity = $_GET["q"];
+        $phonecaseid = $_GET["ps"];
+        $consulta2 = $this->connection->prepare("SELECT value FROM phonecase where id = $phonecaseid");
+        $consulta2->execute();
+        $resultados2 = $consulta2 ->fetchAll();
+        if(count($resultados2) > 0){
+          $valueOfPhoneCase = $resultados2[0]["value"]; 
+          $listShoppingCarts=array();
+
+          $sql = "DELETE FROM shoppingcart WHERE id=$id";
+          $consulta = $this->connection->prepare($sql);
+          $consulta->execute();
+
+          if($consulta == True){
+              $changeQuantity = $valueOfPhoneCase + $quantity;
+              $sql1 = "UPDATE phonecase SET value=$changeQuantity WHERE id=$phonecaseid";
+              $consulta1= $this->connection->prepare($sql1);
+              $consulta1->execute();
+              if($consulta1 == True){
+                echo '<script>alert("Delete thành công")</script>';
+                require_once __DIR__ . "/../view/shoppingCart.php"; 
+              }else{
+                echo '<script>alert("Delete thất bại, vui lòng kiểm tra lại")</script>';
+                require_once __DIR__ . "/../view/shoppingCart.php";
+              }
+          }
+          else{
+              echo '<script>alert("Delete thất bại, vui lòng kiểm tra lại")</script>';
+              require_once __DIR__ . "/../view/shoppingCart.php";
+          }
+
+        }
+        else{
           echo '<script>alert("Delete thất bại, vui lòng kiểm tra lại")</script>';
           require_once __DIR__ . "/../view/shoppingCart.php";
       }
+      }
+      catch (Exception $e) {
+          echo '<script>alert("Có lỗi,' . $e -> getMessage() . '")</script>';
+          require_once __DIR__ . "/../view/shoppingCart.php";
+      }
+      
   }
     public function insert(){
         $phonecaseid = $_POST["phonecaseid"];
